Refactor: Renamed Registerforer to Registrar throughout the application for consistency and English naming
